Add files via upload
Logon works, checking password in an array. Wrong password redirects to a page with a retry button, which redirects to a 2nd logon page.
Calculator works, changed values to symbols for math functions to display on results page. Correct logon will redirect to a calculator function page with the option to choose two numbers and a math function. Will add comments in next commit for future reference.